gcp/automated-threat-response-unified-security-functions/code/terraform/functions/requirements.txt
google-cloud-monitoring==2.21.0
google-cloud-securitycenter==1.36.0
google-cloud-pubsub==2.25.0
google-cloud-logging==3.11.0
google-cloud-compute==1.20.0
google-cloud-resource-manager==1.12.5
functions-framework==3.*